+++ g2p_programs/models/managers/__init__.py
@@ -25,3 +25,4 @@ from . import deduplication_manager
 from . import notification_manager
 from . import program_manager
 from . import cycle_manager
+from . import entitlement_manager